// frontend/src/api/models/convertedBalance.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { MoneyWithSign } from './moneyWithSign';

export interface ConvertedBalance {
  balance: MoneyWithSign;
  rate: number;
  rateDate: string;
}